requirements.txt
python-dispatch==0.2.3
python-dotenv==1.2.2
websocket-client==1.9.0
